+++ CHANGELOG.md
@@ -15,6 +15,10 @@ List all changes after the last release here (newer on top). Each change on a se
 - Discounts: create different admin module for archived discounts to fix breadcrumbs
 - Fix product pagination by not overriding the state with undefined values
 
+### Fixed
+
+- Middleware: fix so it trys to take the users timezone first, then the suppliers, last the projects TIME_ZONE
+
 ### Changed
 
 - Front: customize sort options through settings